refactor EmailManager
