--- application/src/main/java/team/themoment/officialgsm/common/util/EmailUtil.java
@@ -1,5 +1,0 @@
-package team.themoment.officialgsm.common.util;
-
-public interface EmailUtil {
-    String getOauthEmailDomain(String email);
-}
